+++ traffic_ops/traffic_ops_golang/server/servers.go
@@ -93,6 +93,7 @@ func (s *TOServer) Sanitize() {
 
 func (s *TOServer) Validate() error {
 	s.Sanitize()
+	version := s.APIInfo().Version
 	noSpaces := validation.NewStringRule(tovalidate.NoSpaces, "cannot contain spaces")
 
 	errs := []error{}
@@ -106,6 +107,12 @@ func (s *TOServer) Validate() error {
 
 	if s.IP6IsService != nil && *s.IP6IsService && (s.IP6Address == nil || *s.IP6Address == "") {
 		errs = append(errs, tc.EmptyAddressCannotBeAServiceAddressError)
+	}
+
+	if version.Major >= 2 {
+		if (s.IPIsService == nil || !*s.IPIsService) && (s.IP6IsService == nil || !*s.IP6IsService) {
+			errs = append(errs, tc.NeedsAtLeastOneServiceAddressError)
+		}
 	}
 
 	validateErrs := validation.Errors{