+++ CHANGELOG.md
@@ -82,6 +82,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - /parameters/:id/unassigned_profiles
   - /parameters/profile/:name
   - /parameters/validate
+  - /phys_locations/trimmed
   - /phys_locations/:id (GET)
   - /profile/:id (GET)
   - /profile/:id/unassigned_parameters